nixos/nix-channel: only try to remove the nix-channel binary if it exists
It may not be there if `system.disableInstallerTools = true`.

+++ nixos/modules/config/nix-channel.nix
@@ -86,7 +86,9 @@ in
       '';
 
     environment.extraSetup = mkIf (!cfg.channel.enable) ''
-      rm $out/bin/nix-channel
+      if [ -f "$out/bin/nix-channel" ]; then
+        rm $out/bin/nix-channel
+      fi
     '';
 
     # NIX_PATH has a non-empty default according to Nix docs, so we don't unset
